requirements.txt
Flask==3.0.3
python-docx==1.1.2
requests==2.32.3
python-dotenv==0.19.2
PyPDF2==3.0.1
gunicorn>=23.0.0